fix: 依赖循环解决
