+++ spring-boot-polaris-context/src/main/java/cn/polarismesh/boot/context/PolarisContextProperties.java
@@ -17,28 +17,13 @@
 
 package cn.polarismesh.boot.context;
 
-import static cn.polarismesh.boot.context.PolarisContextConst.DEFAULT_LOCAL_HOST;
-
-import com.tencent.polaris.api.config.ConfigProvider;
-import com.tencent.polaris.api.config.Configuration;
-import com.tencent.polaris.factory.ConfigAPIFactory;
-import com.tencent.polaris.factory.config.ConfigurationImpl;
-import java.io.IOException;
-import java.net.Socket;
-import java.util.Collection;
-import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-import org.springframework.util.CollectionUtils;
-import org.springframework.util.StringUtils;
 
 /**
  * Properties for Polaris {@link com.tencent.polaris.client.api.SDKContext}
  *
  * @author Haotian Zhang
  */
-@Component
 @ConfigurationProperties(prefix = PolarisContextConst.PREFIX)
 public class PolarisContextProperties {
 
@@ -47,51 +32,11 @@ public class PolarisContextProperties {
      */
     private String address;
 
-    @Autowired
-    private List<PolarisConfigModifier> modifierList;
-
     public String getAddress() {
         return address;
     }
 
     public void setAddress(String address) {
         this.address = address;
-    }
-
-    protected Configuration configuration() {
-        ConfigurationImpl configuration = (ConfigurationImpl) ConfigAPIFactory
-                .defaultConfig(ConfigProvider.DEFAULT_CONFIG);
-        configuration.setDefault();
-        Collection<PolarisConfigModifier> modifiers = modifierList;
-        if (!CollectionUtils.isEmpty(modifiers)) {
-            for (PolarisConfigModifier modifier : modifiers) {
-                modifier.modify(configuration);
-            }
-        }
-        String defaultHost = getHost(configuration);
-        configuration.getGlobal().getAPI().setBindIP(defaultHost);
-        return configuration;
-    }
-
-    private String getHost(Configuration configuration) {
-        String serverAddress = null;
-        List<String> addresses = configuration.getGlobal().getServerConnector().getAddresses();
-        if (!CollectionUtils.isEmpty(addresses)) {
-            serverAddress = addresses.get(0);
-        }
-        if (!StringUtils.hasText(serverAddress)) {
-            return DEFAULT_LOCAL_HOST;
-        }
-        return getLocalHost(serverAddress);
-    }
-
-    private static String getLocalHost(String serverAddress) {
-        String[] tokens = serverAddress.split(":");
-        try (Socket socket = new Socket(tokens[0], Integer.parseInt(tokens[1]))) {
-            return socket.getLocalAddress().getHostAddress();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return DEFAULT_LOCAL_HOST;
     }
 }
