+++ spring-boot-polaris-context/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,1 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=cn.polarismesh.boot.context.PolarisContextConfiguration
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=cn.polarismesh.boot.context.PolarisContextAutoConfiguration
